[DEV] - Fixed strategy to add matches

+++ src/main/java/ro/unibuc/fmi/karate_management_platform/dtos/competition/match/request/IndividualKataMatchResultRequest.java
@@ -15,5 +15,5 @@ import java.util.Map;
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class IndividualKataMatchResultRequest extends IndividualMatchResultRequest {
-    private Map<Long, Long> athleteScores; // Map<Athlete ID, Score>
+    private Map<Long, Double> athleteScores; // Map<Athlete ID, Score>
 }
