+++ src/main/java/ro/unibuc/fmi/karate_management_platform/dtos/competition/match/request/TeamKataMatchResultRequest.java
@@ -14,5 +14,5 @@ import java.util.Map;
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class TeamKataMatchResultRequest extends TeamMatchResultRequest {
-    private Map<Long, Long> teamScores; // Map<teamId, score>
+    private Map<Long, Double> teamScores; // Map<teamId, score>
 }
